+++ src/ios/BrazePlugin.h
@@ -81,6 +81,9 @@
 - (void)getFeatureFlagBooleanProperty:(CDVInvokedUrlCommand *)command;
 - (void)getFeatureFlagStringProperty:(CDVInvokedUrlCommand *)command;
 - (void)getFeatureFlagNumberProperty:(CDVInvokedUrlCommand *)command;
+- (void)getFeatureFlagTimestampProperty:(CDVInvokedUrlCommand *)command;
+- (void)getFeatureFlagJSONProperty:(CDVInvokedUrlCommand *)command;
+- (void)getFeatureFlagImageProperty:(CDVInvokedUrlCommand *)command;
 - (void)logFeatureFlagImpression:(CDVInvokedUrlCommand *)command;
 
 @end